extract reporting class
